// client/src/objs.tsx
export enum Objetivos {
    Objetivo1 = 'Fin de la pobreza: Erradicar la pobreza extrema y reducir la desigualdad económica.',
    Objetivo2 = 'Hambre cero: Lograr la seguridad alimentaria, la mejora de la nutrición y promover la agricultura sostenible.',
    Objetivo3 = 'Salud y bienestar: Garantizar una vida saludable y promover el bienestar para todos en todas las edades.',
    Objetivo4 = 'Educación de calidad: Asegurar el acceso a una educación inclusiva, equitativa y de calidad.',
    Objetivo5 = 'Igualdad de género: Lograr la igualdad entre los géneros y empoderar a todas las mujeres y niñas.',
    Objetivo6 = 'Agua limpia y saneamiento: Garantizar la disponibilidad y la gestión sostenible del agua y el saneamiento para todos.',
    Objetivo7 = 'Energía asequible y no contaminante: Garantizar el acceso a una energía asequible, fiable, sostenible y moderna para todos.',
    Objetivo8 = 'Trabajo decente y crecimiento económico: Promover un crecimiento económico sostenido, inclusivo y sostenible, el empleo pleno y productivo y el trabajo decente para todos.',
    Objetivo9 = 'Industria, innovación e infraestructura: Construir una infraestructura resiliente, promover la industrialización inclusiva y sostenible y fomentar la innovación.',
    Objetivo10 = 'Reducir las desigualdades: Reducir la desigualdad económica, social y territorial mediante políticas inclusivas y sostenibles.',
    Objetivo11 = 'Ciudades y comunidades sostenibles: Lograr que las ciudades y los asentamientos humanos sean inclusivos, seguros, resilientes y sostenibles.',
    Objetivo12 = 'Producción y consumo responsables: Garantizar patrones de producción y consumo sostenibles y responsables.',
    Objetivo13 = 'Acción por el clima: Adoptar medidas urgentes para combatir el cambio climático y sus efectos.',
    Objetivo14 = 'Vida submarina: Conservar y utilizar de forma sostenible los océanos, los mares y los recursos marinos para el desarrollo sostenible.',
    Objetivo15 = 'Vida de ecosistemas terrestres: Proteger, restaurar y promover la utilización sostenible de los ecosistemas terrestres, gestionar de manera sostenible los bosques, combatir la desertificación, detener y revertir la degradación de la tierra y frenar la pérdida de la diversidad biológica.',
    Objetivo16 = 'Paz, justicia e instituciones sólidas: Promover sociedades pacíficas e inclusivas para el desarrollo sostenible, proporcionar acceso a la justicia para todos y crear instituciones eficaces, responsables e inclusivas a todos los niveles.',
    Objetivo17 = 'Alianzas para lograr los objetivos: Fortalecer los medios de implementación y revitalizar la Alianza Mundial para el Desarrollo Sostenible.'
}
  